Refactor types and data
