+++ src/lib/types/searchTypes.ts
@@ -1,0 +1,9 @@
+export interface TreeNode {
+  name: string;
+  children?: TreeNode[];
+}
+
+export interface Snapshot {
+  activeNode: string;
+  visitedNodes: string[];
+}
